Extract common retry session to independent module.
