+++ bookworm/ocr_engines/baidu_ocr.py
@@ -20,27 +20,9 @@ from .base import (
     OcrNetworkError,
     OcrProcessingError,
 )
+from ._shared import create_session_with_retries
 
 log = logger.getChild(__name__)
-
-
-def create_session_with_retries() -> requests.Session:
-    """
-    Creates a requests session with a robust retry strategy.
-    """
-    session = requests.Session()
-    retries = Retry(
-        total=3,
-        backoff_factor=0.5,
-        status_forcelist=[429, 500, 502, 503, 504],
-        allowed_methods=["POST", "GET"],
-        respect_retry_after_header=True,
-    )
-    adapter = HTTPAdapter(max_retries=retries)
-    session.mount("http://", adapter)
-    session.mount("https://", adapter)
-    return session
-
 
 # Define constants for configuration keys
 BAIDU_API_KEY_NAME = "baidu_api_key"
